+++ src/expressions/visitors/InvalidNode.cpp
@@ -1,18 +1,19 @@
 // Copyright 2007-2026, RTE (https://www.rte-france.com)
 // SPDX-License-Identifier: MPL-2.0
 
+#include <string>
 #include <antares/expressions/visitors/InvalidNode.h>
 
 namespace Antares::Expressions::Visitors
 {
 
 InvalidNode::InvalidNode(const std::string& node_name):
-    std::invalid_argument("Node visitor encountered an invalid node type: " + node_name)
+    std::invalid_argument(std::string("Node visitor encountered an invalid node type: ") + node_name)
 {
 }
 
 NodeTypeShouldBeInExtraOutput::NodeTypeShouldBeInExtraOutput(const std::string& node_name):
-    std::invalid_argument("This type of node: '" + node_name
+    std::invalid_argument(std::string("This type of node: '") + node_name
                           + "' should only be used in extra outputs expressions")
 {
 }