+++ src/solver/optimisation/include/antares/solver/optimisation/adequacy_patch_csr/csr_quadratic_problem.h
@@ -33,6 +33,8 @@ private:
     void setFictitiousLoadConstraints(ConstraintBuilder& builder);
     void setMaxEnsLoadConstraints(ConstraintBuilder& builder);
     void setBindingConstraints(ConstraintBuilder& builder);
+    void setFlowBasedConstraints(ConstraintBuilder& builder);
+    void setGemsConservationConstraint(ConstraintBuilder& builder);
 };
 
 } // namespace Antares::Solver::Optimization